requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.3.1
